% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizacao.r
\name{checa_deps_plot}
\alias{checa_deps_plot}
\title{Checa Dependencias De Visualizacao}
\usage{
checa_deps_plot(tipo = c("3d", "2d"))
}
\arguments{
\item{tipo}{um de \code{c("3d", "2d")} indicando o tipo de grafico desejado}
}
\value{
Caso dependencias existam, retorna NULL; do contrario lanca erro
}
\description{
Checa se pacotes necessarios existem para cada tipo de visualizacao, retornando erro
}
