% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvacolina.r
\docType{package}
\name{curvacolina-package}
\alias{curvacolina}
\alias{curvacolina-package}
\title{Pacote \code{curvacolina}}
\description{
Conjunto De Funcoes Para Processamento De Curvas Colina
}
\author{
\strong{Maintainer}: Lucas Khenayfis \email{lucaskhenayfis@example.com}

}
